develop

+++ src/controllers/demarket/Follow.controller.ts
@@ -16,10 +16,10 @@ class FoLLowController {
             const { followingId, followerId } = request.body;
             if (!followingId && !followerId)
                 return response.status(StatusCodes.BAD_REQUEST).json(new BadRequest("Following id and follower id has been required."));
-            // const accountFollowing = await accountService.findAccountById(followingId);
-            // if (!accountFollowing) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Account is not found."));
-            // const accountFollower = await accountService.findAccountById(followerId);
-            // if (!accountFollower) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Account is not found."));
+            const accountFollowing = await accountService.findAccountById(followingId);
+            if (!accountFollowing) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Account is not found."));
+            const accountFollower = await accountService.findAccountById(followerId);
+            if (!accountFollower) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Account is not found."));
             const existAccountFollow = await followService.checkFollowExistAccount({ followerId: followerId, followingId: followingId });
             if (existAccountFollow) return response.status(StatusCodes.BAD_REQUEST).json({ message: "Follow account has already exist." });
             await followService.createFollowAccount({ followerId: followerId, followingId: followingId });
@@ -40,12 +40,12 @@ class FoLLowController {
             const { followingId, followerId } = request.body;
             if (!followingId && !followerId)
                 return response.status(StatusCodes.BAD_REQUEST).json(new BadRequest("Following id and follower id has been required."));
-            // const accountFollowing = await accountService.findAccountById(followingId);
-            // if (!accountFollowing) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Account is not found."));
-            // const accountFollower = await accountService.findAccountById(followerId);
-            // if (!accountFollower) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Account is not found."));
-            // const existAccountFollow = await followService.checkFollowExistAccount({ followerId: followerId, followingId: followingId });
-            // if (!existAccountFollow) return response.status(StatusCodes.BAD_REQUEST).json({ message: "Follow account is not found." });
+            const accountFollowing = await accountService.findAccountById(followingId);
+            if (!accountFollowing) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Account is not found."));
+            const accountFollower = await accountService.findAccountById(followerId);
+            if (!accountFollower) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Account is not found."));
+            const existAccountFollow = await followService.checkFollowExistAccount({ followerId: followerId, followingId: followingId });
+            if (!existAccountFollow) return response.status(StatusCodes.BAD_REQUEST).json({ message: "Follow account is not found." });
             await followService.deleteFollowAccount({ followerId: followerId, followingId: followingId });
             response.status(StatusCodes.OK).json({ message: "Unfollow account successfully." });
         } catch (error) {
